+++ packages/asset-swapper/src/utils/market_operation_utils/index.ts
@@ -158,6 +158,8 @@ export class MarketOperationUtils {
 
         // Call the sampler contract.
         const samplerPromise = this._sampler.executeAsync(
+            this._sampler.getBlockNumber(),
+            this._sampler.getGasLeft(),
             this._sampler.getTokenDecimals([makerToken, takerToken]),
             // Get native order fillable amounts.
             this._sampler.getLimitOrderFillableTakerAmounts(nativeOrders, this.contractAddresses.exchangeProxy),
@@ -184,6 +186,7 @@ export class MarketOperationUtils {
                 takerAmount,
             ),
             this._sampler.isAddressContract(txOrigin),
+            this._sampler.getGasLeft(),
         );
 
         // Refresh the cached pools asynchronously if required
@@ -191,6 +194,8 @@ export class MarketOperationUtils {
 
         const [
             [
+                blockNumber,
+                gasBefore,
                 tokenDecimals,
                 orderFillableTakerAmounts,
                 outputAmountPerEth,
@@ -198,8 +203,13 @@ export class MarketOperationUtils {
                 dexQuotes,
                 rawTwoHopQuotes,
                 isTxOriginContract,
+                gasAfter,
             ],
         ] = await Promise.all([samplerPromise]);
+
+        // Log the gas metrics
+        _opts.samplerMetrics?.logGasDetails({ gasBefore, gasAfter });
+        _opts.samplerMetrics?.logBlockNumber(blockNumber);
 
         // Filter out any invalid two hop quotes where we couldn't find a route
         const twoHopQuotes = rawTwoHopQuotes.filter(