+++ packages/asset-swapper/test/wrappers.ts
@@ -8,6 +8,7 @@ export * from '../test/generated-wrappers/balance_checker';
 export * from '../test/generated-wrappers/balancer_sampler';
 export * from '../test/generated-wrappers/balancer_v2_sampler';
 export * from '../test/generated-wrappers/bancor_sampler';
+export * from '../test/generated-wrappers/compound_sampler';
 export * from '../test/generated-wrappers/curve_sampler';
 export * from '../test/generated-wrappers/d_o_d_o_sampler';
 export * from '../test/generated-wrappers/d_o_d_o_v2_sampler';